Finish labeling with extended XYZ

+++ deepmd_iterative/labeling/check.py
@@ -6,8 +6,9 @@
 #   SPDX-License-Identifier: AGPL-3.0-only                                                           #
 #----------------------------------------------------------------------------------------------------#
 Created: 2022/01/01
-Last modified: 2024/03/14
+Last modified: 2024/03/31
 """
+
 # Standard library modules
 import logging
 import re
@@ -15,18 +16,9 @@ import sys
 from pathlib import Path
 
 # Local imports
-from deepmd_iterative.common.json import (
-    load_json_file,
-    write_json_file,
-)
-from deepmd_iterative.common.list import (
-    textfile_to_string_list,
-    string_list_to_textfile,
-)
-from deepmd_iterative.common.filesystem import (
-    remove_file,
-    remove_files_matching_glob,
-)
+from deepmd_iterative.common.json import load_json_file, write_json_file
+from deepmd_iterative.common.list import textfile_to_string_list, string_list_to_textfile
+from deepmd_iterative.common.filesystem import remove_file
 from deepmd_iterative.common.check import validate_step_folder
 
 
@@ -42,9 +34,7 @@ def main(
     training_path = current_path.parent
 
     # Log the step and phase of the program
-    logging.info(
-        f"Step: {current_step.capitalize()} - Phase: {current_phase.capitalize()}."
-    )
+    logging.info(f"Step: {current_step.capitalize()} - Phase: {current_phase.capitalize()}.")
     logging.debug(f"Current path :{current_path}")
     logging.debug(f"Training path: {training_path}")
     logging.debug(f"Program path: {deepmd_iterative_path}")
@@ -66,17 +56,13 @@ def main(
     if curr_iter > 1:
         prev_iter = curr_iter - 1
         padded_prev_iter = str(prev_iter).zfill(3)
-        previous_labeling_json = load_json_file(
-            (control_path / f"labeling_{padded_prev_iter}.json")
-        )
+        previous_labeling_json = load_json_file((control_path / f"labeling_{padded_prev_iter}.json"))
         del prev_iter, padded_prev_iter
     else:
         previous_labeling_json = {}
 
-
-    labeling_program = labeling_json["labeling_json"]
+    labeling_program = labeling_json["labeling_program"]
     logging.debug(f"labeling_program: {labeling_program}")
-    labeling_program_up = labeling_program.upper()
 
     # Check if we can continue
     if not labeling_json["is_launched"]:
@@ -91,21 +77,13 @@ def main(
     candidates_step_count = {0: 0, 1: 0}
 
     for system_auto_index, system_auto in enumerate(labeling_json["systems_auto"]):
-        logging.info(
-            f"Processing system: {system_auto} ({system_auto_index + 1}/{len(labeling_json['systems_auto'])})"
-        )
+        logging.info(f"Processing system: {system_auto} ({system_auto_index + 1}/{len(labeling_json['systems_auto'])})")
         system_path = current_path / system_auto
 
-        system_candidates_count = labeling_json["systems_auto"][system_auto][
-            "candidates_count"
-        ]
-        system_disturbed_candidates_count = labeling_json["systems_auto"][system_auto][
-            "disturbed_candidates_count"
-        ]
+        system_candidates_count = labeling_json["systems_auto"][system_auto]["candidates_count"]
+        system_disturbed_candidates_count = labeling_json["systems_auto"][system_auto]["disturbed_candidates_count"]
 
-        candidates_expected_count += (
-            system_candidates_count + system_disturbed_candidates_count
-        )
+        candidates_expected_count += system_candidates_count + system_disturbed_candidates_count
 
         # One step, if it is skipped, we don't care if one is not converged or failed
         system_candidates_skipped_count = 0
@@ -120,21 +98,15 @@ def main(
         system_candidates_not_converged = {0: [], 1: []}
         system_candidates_failed = {0: [], 1: []}
 
-        logging.debug(
-            f"system_candidates_count + system_disturbed_candidates_count: {system_candidates_count + system_disturbed_candidates_count}"
-        )
+        logging.debug(f"system_candidates_count + system_disturbed_candidates_count: {system_candidates_count + system_disturbed_candidates_count}")
         if system_candidates_count + system_disturbed_candidates_count == 0:
             # TODO Because no candidates, we "fake-fill" with previous labeling values for the timings
             if curr_iter > 1:
-                labeling_json["systems_auto"][system_auto][
-                    "timings_s"
-                ] = previous_labeling_json["systems_auto"][system_auto]["timings_s"]
+                labeling_json["systems_auto"][system_auto]["timings_s"] = previous_labeling_json["systems_auto"][system_auto]["timings_s"]
             else:
                 labeling_json["systems_auto"][system_auto]["timings_s"] = [1800, 3600]
             labeling_json["systems_auto"][system_auto]["candidates_skipped_count"] = 0
-            labeling_json["systems_auto"][system_auto][
-                "disturbed_candidates_skipped_count"
-            ] = 0
+            labeling_json["systems_auto"][system_auto]["disturbed_candidates_skipped_count"] = 0
             continue
 
         # TODO Use a function to parse CP2K output
@@ -152,7 +124,7 @@ def main(
                     system_output_cp2k_file = {}
                     system_output_cp2k = {}
                     for step in [0, 1]:
-                        system_output_cp2k_file[step] = (labeling_step_path / f"{step+1}_labeling_{padded_labeling_step}.out")
+                        system_output_cp2k_file[step] = labeling_step_path / f"{step+1}_labeling_{padded_labeling_step}.out"
                         if system_output_cp2k_file[step].is_file():
                             system_output_cp2k[step] = textfile_to_string_list(system_output_cp2k_file[step])
 
@@ -160,7 +132,7 @@ def main(
                                 candidates_step_count[step] += 1
                                 system_candidates_converged_count[step] += 1
                                 if any("T I M I N G" in _ for _ in system_output_cp2k[step]):
-                                    system_timings[step] = [_ for _ in system_output_cp2k[step] if "CP2K                                 1  1.0" in _ ]
+                                    system_timings[step] = [_ for _ in system_output_cp2k[step] if "CP2K                                 1  1.0" in _]
                                     system_tinings_sum[step] += float(system_timings[step][0].split(" ")[-1])
 
                             elif any("SCF run NOT converged" in _ for _ in system_output_cp2k[step]):
@@ -173,7 +145,7 @@ def main(
                     del system_output_cp2k_file, system_output_cp2k
 
                 elif labeling_program == "orca":
-                    system_output_orca_file = (labeling_step_path / f"1_labeling_{padded_labeling_step}.out")
+                    system_output_orca_file = labeling_step_path / f"1_labeling_{padded_labeling_step}.out"
                     if system_output_orca_file.is_file():
                         system_output_orca = textfile_to_string_list(system_output_orca_file)
                         if any("ORCA TERMINATED NORMALLY" in _ for _ in system_output_orca) and any("SCF CONVERGED" in _ for _ in system_output_orca):
@@ -181,7 +153,7 @@ def main(
                             system_candidates_converged_count[0] += 1
                             if any("Sum of individual times" in _ for _ in system_output_orca):
                                 system_timings[0] = [_ for _ in system_output_orca if "Sum of individual times" in _]
-                                pattern = r'(\d+\.\d+) sec'
+                                pattern = r"(\d+\.\d+) sec"
                                 matches = re.search(pattern, system_timings)
                                 if matches:
                                     system_tinings_sum[0] += float(matches.group(1))
@@ -191,7 +163,7 @@ def main(
                         system_candidates_failed[0].append(f"{system_output_orca_file}")
 
         if system_disturbed_candidates_count > 0:
-            for labeling_step in range(system_candidates_count,system_candidates_count + system_disturbed_candidates_count):
+            for labeling_step in range(system_candidates_count, system_candidates_count + system_disturbed_candidates_count):
                 padded_labeling_step = str(labeling_step).zfill(5)
                 labeling_step_path = system_path / padded_labeling_step
 
@@ -205,17 +177,17 @@ def main(
                         system_output_cp2k_file = {}
                         system_output_cp2k = {}
                         for step in [0, 1]:
-                            system_output_cp2k_file[step] = (labeling_step_path / f"{step+1}_labeling_{padded_labeling_step}.out")
+                            system_output_cp2k_file[step] = labeling_step_path / f"{step+1}_labeling_{padded_labeling_step}.out"
                             if system_output_cp2k_file[step].is_file():
                                 system_output_cp2k[step] = textfile_to_string_list(system_output_cp2k_file[step])
                                 if any("SCF run converged in" in _ for _ in system_output_cp2k[step]):
                                     candidates_step_count[step] += 1
                                     system_candidates_converged_count[step] += 1
                                     if any("T I M I N G" in _ for _ in system_output_cp2k[step]):
-                                        system_timings[step] = [_ for _ in system_output_cp2k[step] if "CP2K                                 1  1.0" in _ ]
+                                        system_timings[step] = [_ for _ in system_output_cp2k[step] if "CP2K                                 1  1.0" in _]
                                         system_tinings_sum[step] += float(system_timings[step][0].split(" ")[-1])
-                                elif any("SCF run NOT converged" in _ for _ in system_output_cp2k[step] ):
-                                    system_candidates_not_converged[step].append( f"{system_output_cp2k_file[step]}")
+                                elif any("SCF run NOT converged" in _ for _ in system_output_cp2k[step]):
+                                    system_candidates_not_converged[step].append(f"{system_output_cp2k_file[step]}")
                                 else:
                                     system_candidates_failed[0].append(f"{system_output_orca_file}")
                             else:
@@ -223,7 +195,7 @@ def main(
                         del system_output_cp2k_file, system_output_cp2k
 
                     elif labeling_program == "orca":
-                        system_output_orca_file = (labeling_step_path / f"1_labeling_{padded_labeling_step}.out")
+                        system_output_orca_file = labeling_step_path / f"1_labeling_{padded_labeling_step}.out"
                         if system_output_orca_file.is_file():
                             system_output_orca = textfile_to_string_list(system_output_orca_file)
                             if any("ORCA TERMINATED NORMALLY" in _ for _ in system_output_orca) and any("SCF CONVERGED" in _ for _ in system_output_orca):
@@ -231,7 +203,7 @@ def main(
                                 system_candidates_converged_count[0] += 1
                                 if any("Sum of individual times" in _ for _ in system_output_orca):
                                     system_timings[0] = [_ for _ in system_output_orca if "Sum of individual times" in _]
-                                    pattern = r'(\d+\.\d+) sec'
+                                    pattern = r"(\d+\.\d+) sec"
                                     matches = re.search(pattern, system_timings)
                                     if matches:
                                         system_tinings_sum[0] += float(matches.group(1))
@@ -257,69 +229,44 @@ def main(
         # For the very special case where there are no converged subsystems (e.g., if you skipped all jobs)
         for step, default_timing in enumerate([900.0, 3600.0]):
             if system_candidates_converged_count[step] != 0:
-                timings[step] = (
-                    system_tinings_sum[step] / system_candidates_converged_count[step]
-                )
+                timings[step] = system_tinings_sum[step] / system_candidates_converged_count[step]
             else:
                 timings[step] = default_timing
         del step, default_timing
         del system_tinings_sum, system_candidates_converged_count, system_timings
 
-        labeling_json["systems_auto"][system_auto]["timings_s"] = [
-            timings[0],
-            timings[1],
-        ]
-        labeling_json["systems_auto"][system_auto][
-            "candidates_skipped_count"
-        ] = system_candidates_skipped_count
-        labeling_json["systems_auto"][system_auto][
-            "disturbed_candidates_skipped_count"
-        ] = system_disturbed_candidates_skipped_count
+        labeling_json["systems_auto"][system_auto]["timings_s"] = [timings[0], timings[1]]
+        labeling_json["systems_auto"][system_auto]["candidates_skipped_count"] = system_candidates_skipped_count
+        labeling_json["systems_auto"][system_auto]["disturbed_candidates_skipped_count"] = system_disturbed_candidates_skipped_count
         del timings
 
         for step in [0, 1]:
-            not_converged_file = (
-                system_path / f"{system_auto}_step{step+1}_not_converged.txt"
-            )
+            if labeling_program == "orca" and step == 1:
+                continue
+            not_converged_file = system_path / f"{system_auto}_step{step+1}_not_converged.txt"
             remove_file(not_converged_file)
             if len(system_candidates_not_converged[step]) > 0:
-                string_list_to_textfile(
-                    not_converged_file, system_candidates_not_converged[step]
-                )
-                logging.warning(
-                    f"{system_auto} | step {step+1}: {len(system_candidates_not_converged[step])} jobs did not converge. List in '{not_converged_file}'"
-                )
+                string_list_to_textfile(not_converged_file, system_candidates_not_converged[step])
+                logging.warning(f"{system_auto} | step {step+1}: {len(system_candidates_not_converged[step])} jobs did not converge. List in '{not_converged_file}'")
             del not_converged_file
 
             failed_file = system_path / f"{system_auto}_step{step+1}_failed.txt"
             remove_file(failed_file)
             if len(system_candidates_failed[step]) > 0:
                 string_list_to_textfile(failed_file, system_candidates_failed[step])
-                logging.warning(
-                    f"{system_auto} | step {step+1}: {len(system_candidates_failed[step])} jobs did not converge. List in '{failed_file}'."
-                )
+                logging.warning(f"{system_auto} | step {step+1}: {len(system_candidates_failed[step])} jobs did not converge. List in '{failed_file}'.")
             del failed_file
 
         del step
         del system_candidates_not_converged, system_candidates_failed
 
         remove_file(system_path / f"{system_auto}_skipped.txt")
-        if (
-            system_candidates_skipped_count > 0
-            or system_disturbed_candidates_skipped_count > 0
-        ):
+        if system_candidates_skipped_count > 0 or system_disturbed_candidates_skipped_count > 0:
             skipped_file = system_path / f"{system_auto}_skipped.txt"
-            system_candidates_skipped_total = (
-                system_candidates_skipped + system_disturbed_candidates_skipped
-            )
-            system_candidates_skipped_total_count = (
-                system_candidates_skipped_count
-                + system_disturbed_candidates_skipped_count
-            )
+            system_candidates_skipped_total = system_candidates_skipped + system_disturbed_candidates_skipped
+            system_candidates_skipped_total_count = system_candidates_skipped_count + system_disturbed_candidates_skipped_count
             string_list_to_textfile(skipped_file, system_candidates_skipped_total)
-            logging.info(
-                f"{system_auto}: {system_candidates_skipped_total_count} jobs skipped ({system_candidates_skipped_count}|{system_disturbed_candidates_skipped_count}). List in '{skipped_file}'."
-            )
+            logging.info(f"{system_auto}: {system_candidates_skipped_total_count} jobs skipped ({system_candidates_skipped_count}|{system_disturbed_candidates_skipped_count}). List in '{skipped_file}'.")
 
             del system_candidates_skipped_total, system_candidates_skipped_total_count
             del skipped_file
@@ -329,15 +276,10 @@ def main(
             )
             del system_candidates_skipped, system_disturbed_candidates_skipped
 
-        del (
-            system_candidates_count,
-            system_disturbed_candidates_count,
-        )
+        del system_candidates_count, system_disturbed_candidates_count
         del system_path, labeling_step, labeling_step_path, padded_labeling_step
 
-        logging.info(
-            f"Processed system: {system_auto} ({system_auto_index + 1}/{len(labeling_json['systems_auto'])})"
-        )
+        logging.info(f"Processed system: {system_auto} ({system_auto_index + 1}/{len(labeling_json['systems_auto'])})")
     del system_auto, system_auto_index
 
     # Check first step, write warning if not converged/failed and CP2K, abort if not converged/failed and ORCA
@@ -357,9 +299,7 @@ def main(
 
     logging.info(f"-" * 88)
     # Update the booleans in the exploration JSON
-    if candidates_expected_count == (
-        candidates_step_count[1] + candidates_skipped_count
-    ):
+    if candidates_expected_count == (candidates_step_count[1] + candidates_skipped_count):
         labeling_json["is_checked"] = True
     del candidates_expected_count, candidates_skipped_count, candidates_step_count
 
@@ -368,9 +308,7 @@ def main(
 
     # End
     logging.info(f"-" * 88)
-    logging.info(
-        f"Step: {current_step.capitalize()} - Phase: {current_phase.capitalize()} is a success!"
-    )
+    logging.info(f"Step: {current_step.capitalize()} - Phase: {current_phase.capitalize()} is a success!")
 
     # Cleaning
     del current_path, control_path, training_path
